==> PART-1_Data-Acquisition.ipynb <==
# Spotify Streaming History Dashboard
## Part 1: Acquiring Data and Track URIs
Summary of process:
- Create dataframe from streaming history json files
- Clean dataframe 
- Populate dataframe with track URIs from library json file
- Populate remaing track URIs from Spotify Web API
***
### Install and Import Required Libraries

!pip install spotipy
!pip install requests
!pip install pandas
!pip install tqdm

import pandas as pd
import numpy as np
import requests
import json
import random
from tqdm.notebook import tqdm_notebook
import time
import spotipy
from spotipy.oauth2 import SpotifyOAuth

### Compile Database of Streaming History

# Put each streaming history json file into pd dataframe
df_file0 = pd.read_json('StreamingHistory0.json')
df_file1 = pd.read_json('StreamingHistory1.json')
df_file2 = pd.read_json('StreamingHistory2.json')
df_file3 = pd.read_json('StreamingHistory3.json')
df_file4 = pd.read_json('StreamingHistory4.json')

# Combine streaming history into one dataframe
df_streams = pd.concat([df_file0, df_file1, df_file2, df_file3, df_file4]).reset_index(drop=True)

# Put library json file into pd dataframe
with open('YourLibrary.json') as json_data:
    library = json.load(json_data)
df_library = pd.DataFrame(library['tracks'])

df_streams.tail()

df_library.head()

### Clean Dataframe and Prep for API

# Drop entries with unknown artist or unknown track
df_streams.drop(df_streams[df_streams.artistName=='Unknown Artist'].index, inplace=True)
df_streams.drop(df_streams[df_streams.trackName=='Unknown Track'].index, inplace=True)
df_streams.reset_index(drop=True, inplace=True)
df_streams.shape

# Remove songs with 0ms listening time
df_streams.drop(df_streams[df_streams.msPlayed==0].index, inplace=True)
df_streams.reset_index(inplace=True, drop=True)

# Remove extra days (we want exactly one year)
df_streams.drop(df_streams[df_streams.endTime<'2021-05-21'].index, inplace=True)
df_streams.drop(df_streams[df_streams.endTime>'2022-05-21'].index, inplace=True)
df_streams.reset_index(inplace=True, drop=True)

# Add URI column 
df_streams['uri'] = np.nan

df_streams.tail()

# Remove "spotify:track:" from URI in the library dataframe
df_library['uri'] = df_library["uri"].str.split(":", expand = True)[2]
df_library.head()

# Populate URI column in streams df with URIs from library column
for s_i, s_row in tqdm_notebook(df_streams.iterrows()):
    for l_i, l_row in df_library.iterrows():
        if ((str(s_row['artistName']) == str(l_row['artist'])) and (str(s_row['trackName']) == str(l_row['track']))):
            df_streams['uri'][s_i] = df_library['uri'][l_i]        
            
df_streams.tail(15)

# Save to CSV
df_streams.to_csv('StreamingHistory_PreURI.csv', index=False)

# Check number of requests required for API
df_streams['uri'].isnull().sum()

### Retrieve Missing URIs from Spotify Web API

df=pd.read_csv('StreamingHistory_PreURI.csv')

# Setup access for Spotify API
client_id = 'ebd35cded9324038b57d33aa816040fe'
client_secret = '4d5c594f164e475e89cb8860506c3a3e'
redirect_uri = 'http://localhost:5000/callback'
scope = 'user-read-recently-played'

sp = spotipy.Spotify(auth_manager=SpotifyOAuth(client_id=client_id,
                                               client_secret=client_secret,
                                               redirect_uri=redirect_uri,
                                               scope=scope))

# Split df in four due to avoid timeout
df1 = df.iloc[0:11000]
df2 = df.iloc[11000:22000]
df3 = df.iloc[22000:33000]
df4 = df.iloc[33000:]


# Request URIs for first quarter
trackURI = []

for index, row in tqdm_notebook(df1.iterrows()):
    artist = row['artistName']
    track = row['trackName']
    uri = row['uri']
    if index % 200 == 0:
        time.sleep(random.uniform(3, 6))
    if pd.isnull(uri):
        query = sp.search(q='artist:' + artist + ' track:' + track, type='track')
        try:
            track_id = query['tracks']['items'][0]['id']
            trackURI.append(track_id)
        except:
            trackURI.append(np.nan)
    else:
        trackURI.append(uri)

# Request URIs for second quarter
for index, row in tqdm_notebook(df2.iterrows()):
    artist = row['artistName']
    track = row['trackName']
    uri = row['uri']
    if index % 200 == 0:
        time.sleep(random.uniform(3, 6))
    if pd.isnull(uri):
        query = sp.search(q='artist:' + artist + ' track:' + track, type='track')
        try:
            track_id = query['tracks']['items'][0]['id']
            trackURI.append(track_id)
        except:
            trackURI.append(np.nan)
    else:
        trackURI.append(uri)

# Request URIs for third quarter
for index, row in tqdm_notebook(df3.iterrows()):
    artist = row['artistName']
    track = row['trackName']
    uri = row['uri']
    if index % 200 == 0:
        time.sleep(random.uniform(3, 6))
    if pd.isnull(uri):
        query = sp.search(q='artist:' + artist + ' track:' + track, type='track')
        try:
            track_id = query['tracks']['items'][0]['id']
            trackURI.append(track_id)
        except:
            trackURI.append(np.nan)
    else:
        trackURI.append(uri)

# Request URIs for fourth quarter
for index, row in tqdm_notebook(df4.iterrows()):
    artist = row['artistName']
    track = row['trackName']
    uri = row['uri']
    if index % 200 == 0:
        time.sleep(random.uniform(3, 6))
    if pd.isnull(uri):
        query = sp.search(q='artist:' + artist + ' track:' + track, type='track')
        try:
            track_id = query['tracks']['items'][0]['id']
            trackURI.append(track_id)
        except:
            trackURI.append(np.nan)
    else:
        trackURI.append(uri)

# Update URI column
df['uri'] = trackURI
df.tail(15)

# Remove entries with no URI
df.dropna(subset=['uri'], inplace=True)
df.reset_index(inplace=True, drop=True)
df = df.astype({"uri": str})

### Save to CSV

df.to_csv('StreamingHistory_PostURI.csv', index=False)
